requirements.txt
fastapi
uvicorn
requests
pydub
miniaudio
aiohttp
sherpa_onnx
numpy
huggingface_hub
opencc-python-reimplemented
python-dateutil
llama-cpp-python @ https://huggingface.co/Luigi/llama-cpp-python-wheels-hf-spaces-free-cpu/resolve/main/llama_cpp_python-0.3.16-cp310-cp310-linux_x86_64.whl